Create Co curricular activities
The program is to be used in assisting a new student (s) in selecting co-curricular activities. The
available co-curricular activities can be categorized in two groups: sporting activities and club/
societies activities. The program will be used to determine the activity/ activities the students will
be undertaking, based on the rules governing the allocation of the activities.